added OsvProcessor
